Double -> float implicit conversion issues fixed

+++ linmath.h
@@ -591,16 +591,16 @@ LINMATH_H_FUNC void mat4x4_arcball(mat4x4 R, mat4x4 M, vec2 _a, vec2 _b, float s
 
 	if (vec2_len(a) < 1.)
 	{
-		z_a = sqrtf(1. - vec2_mul_inner(a, a));
+		z_a = sqrtf(1.f - vec2_mul_inner(a, a));
 	}
 	else
 	{
 		vec2_norm(a, a);
 	}
 
-	if (vec2_len(b) < 1.)
+	if (vec2_len(b) < 1.f)
 	{
-		z_b = sqrtf(1. - vec2_mul_inner(b, b));
+		z_b = sqrtf(1.f - vec2_mul_inner(b, b));
 	}
 	else
 	{
@@ -613,7 +613,7 @@ LINMATH_H_FUNC void mat4x4_arcball(mat4x4 R, mat4x4 M, vec2 _a, vec2 _b, float s
 	vec3 c_;
 	vec3_mul_cross(c_, a_, b_);
 
-	float angle = acos(vec3_mul_inner(a_, b_)) * s;
+	float angle = acosf(vec3_mul_inner(a_, b_)) * s;
 	mat4x4_rotate(R, M, c_[0], c_[1], c_[2], angle);
 }
 #endif // LINMATH
